Add retry on get requests by default and improve error messages
