+++ src/commands/aem/rde/delete.js
@@ -21,6 +21,8 @@ const { loadUpdateHistory } = require('../../../lib/rde-utils');
 const { loadAllArtifacts, groupArtifacts } = require('../../../lib/rde-utils');
 const spinner = require('ora')();
 const { codes: deploymentErrorCodes } = require('../../../lib/deployment-errors');
+const { codes: internalCodes } = require('../../../lib/internal-errors');
+const { AioError } = require('../../../lib/errors');
 
 class DeleteCommand extends BaseCommand {
   async run() {
@@ -68,6 +70,9 @@ class DeleteCommand extends BaseCommand {
       }
     } catch (err) {
       spinner.stop();
+      if (err instanceof AioError) {
+        throw err;
+      }
       throw new internalCodes.INTERNAL_DELETE_ERROR({ messageValues: err });
     }
   }
