+++ src/commands/aem/rde/inspect/inventory.js
@@ -18,6 +18,7 @@ const {
   inspectCommonFlags,
 } = require('../../../../lib/inspect-base-command');
 const { codes: internalCodes } = require('../../../../lib/internal-errors');
+const { AioError } = require('../../../../lib/errors');
 
 class InventoryCommand extends InspectBaseCommand {
   async run() {
@@ -56,6 +57,9 @@ class InventoryCommand extends InspectBaseCommand {
         }
       }
     } catch (err) {
+      if (err instanceof AioError) {
+        throw err;
+      }
       throw new internalCodes.INTERNAL_INVENTORY_ERROR({ messageValues: err });
     }
   }
