+++ src/commands/aem/rde/inspect/osgi-bundles.js
@@ -17,6 +17,7 @@ const {
   inspectCommonFlags,
 } = require('../../../../lib/inspect-base-command');
 const { codes: internalCodes } = require('../../../../lib/internal-errors');
+const { AioError } = require('../../../../lib/errors');
 
 class OsgiBundlesCommand extends InspectBaseCommand {
   async run() {
@@ -56,6 +57,9 @@ class OsgiBundlesCommand extends InspectBaseCommand {
         }
       }
     } catch (err) {
+      if (err instanceof AioError) {
+        throw err;
+      }
       throw new internalCodes.INTERNAL_GET_OSGI_BUNDLES_ERROR({ messageValues: err });
     }
   }
