+++ src/commands/aem/rde/inspect/osgi-services.js
@@ -17,6 +17,7 @@ const {
   inspectCommonFlags,
 } = require('../../../../lib/inspect-base-command');
 const { codes: internalCodes } = require('../../../../lib/internal-errors');
+const { AioError } = require('../../../../lib/errors');
 
 class OsgiServicesCommand extends InspectBaseCommand {
   async run() {
@@ -57,6 +58,9 @@ class OsgiServicesCommand extends InspectBaseCommand {
         }
       }
     } catch (err) {
+      if (err instanceof AioError) {
+        throw err;
+      }
       throw new internalCodes.INTERNAL_GET_OSGI_SERVICES_ERROR({ messageValues: err });
     }
   }
