+++ src/commands/aem/rde/inspect/request-logs/index.js
@@ -18,6 +18,7 @@ const {
   inspectCommonFlags,
 } = require('../../../../../lib/inspect-base-command');
 const { codes: internalCodes } = require('../../../../../lib/internal-errors');
+const { AioError } = require('../../../../../lib/errors');
 
 class RequestLogsCommand extends InspectBaseCommand {
   async run() {
@@ -56,6 +57,9 @@ class RequestLogsCommand extends InspectBaseCommand {
         }
       }
     } catch (err) {
+      if (err instanceof AioError) {
+        throw err;
+      }
       throw new internalCodes.INTERNAL_REQUEST_LOGS_ERROR({ messageValues: err });
     }
   }
