+++ src/commands/aem/rde/restart.js
@@ -12,6 +12,8 @@
 'use strict';
 
 const { BaseCommand, cli } = require('../../../lib/base-command');
+const { codes: internalCodes } = require('../../../lib/internal-errors');
+const { AioError } = require('../../../lib/errors');
 const spinner = require('ora')();
 
 class RestartCommand extends BaseCommand {
@@ -24,6 +26,9 @@ class RestartCommand extends BaseCommand {
       cli.log(`Environment restarted.`);
     } catch (err) {
       spinner.stop();
+      if (err instanceof AioError) {
+        throw err;
+      }
       throw new internalCodes.INTERNAL_RESTART_ERROR({ messageValues: err });
     }
   }
