+++ src/commands/aem/rde/status.js
@@ -14,6 +14,7 @@
 const { BaseCommand, cli, Flags } = require('../../../lib/base-command');
 const { loadAllArtifacts, groupArtifacts } = require('../../../lib/rde-utils');
 const { codes: internalCodes } = require('../../../lib/internal-errors');
+const { AioError } = require('../../../lib/errors');
 const spinner = require('ora')();
 
 class StatusCommand extends BaseCommand {
@@ -63,6 +64,9 @@ class StatusCommand extends BaseCommand {
       );
     } catch (err) {
       spinner.stop();
+      if (err instanceof AioError) {
+        throw err;
+      }
       throw new internalCodes.INTERNAL_STATUS_ERROR({ messageValues: err });
     }
   }
